💄 Style: styled파일 및 index.js 파일 추가
